cf_load_parquet_to_bq/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudevents==1.2.0
db-dtypes==1.0.1
decorator==5.1.1
deprecation==2.1.0
Flask==2.1.2
frozenlist==1.3.0
fsspec==2022.5.0
functions-framework==3.0.0
gcsfs==2022.5.0
google-api-core==2.7.3
google-auth==2.6.6
google-auth-oauthlib==0.5.1
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-bigquery==3.1.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-logging==3.1.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-gbq==0.17.5
proto-plus==1.20.3
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
utils==1.0.1
watchdog==1.0.2
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
